feat: add buttons and manipulate the DOM with event listeners
We create three buttons, one for each selection. Add an event listener
to the buttons that call our playRound function with the correct
userChoice every time a button is clicked. Add a div for displaying
results. Display the running score, and announce a winner of the game
once one player reaches 5 points.